Current state of exporter.
Not perfect, still struggling to find a good way to terminate the listen
thread for the web socket when it becomes stale as it seems to do quite
often. Current plan is to abort main loop and recycle exporter. Should
find a better overall solution to this, but I suck to much at the
asyncio parts.